requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cairocffi==1.7.1
CairoSVG==2.8.2
certifi==2025.4.26
cffi==1.17.1
cssselect2==0.8.0
defusedxml==0.7.1
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.79.0
pillow==11.2.1
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
sniffio==1.3.1
tinycss2==1.4.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
webencodings==0.5.1
